refactor: PlusMenu 컴포넌트 분리 및 호버 영역 변경
